refactor: extract MovieList component and fetch data with useQuery
